server/requirements.txt
# Django and REST Framework
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.3.1
django-filter==24.1
django-phonenumber-field==8.1.0
phonenumbers==9.0.9
Pillow==10.2.0

# JWT Authentication
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0

# Database
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# Environment variables
python-decouple==3.8

# API Documentation
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pytz==2025.2
uritemplate==4.2.0

# Security
cryptography==45.0.5
cffi==1.17.1
pycparser==2.22

# Server
gunicorn==23.0.0
uvicorn==0.29.0

# Caching
django-redis==5.4.0
redis==5.0.1

# Testing
pytest==8.4.1
pytest-django==4.8.0

# Utilities
requests==2.32.4
urllib3==2.5.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10

